bug fix at Play.java
